services/inventory-service/requirements.txt
flask
prometheus-flask-exporter==0.22.3
opentelemetry-api==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-exporter-jaeger==1.15.0
opentelemetry-instrumentation-flask==0.36b0
requests==2.28.2
werkzeug 